requirements.txt
pydantic-ai
PyPDF2
python-dotenv
chromadb
openai
pyyaml
gradio
markdown
